backend/requirements.txt
# requirements.txt

aiofiles==23.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asgiref==3.9.2
asttokens==3.0.0
attrs==25.3.0
auth0-python==4.13.0
backoff==2.2.1
bcrypt==4.3.0
beanie==1.26.0
billiard==4.2.2
black==24.4.2
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
celery==5.3.6
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chromadb==1.1.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.130.0
cryptography==46.0.1
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.112.0
filelock==3.19.1
flatbuffers==25.9.23
flower==2.0.1
frozenlist==1.7.0
fsspec==2025.9.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.75.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.35.1
humanfriendly==10.0
humanize==4.13.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
instructor==1.11.3
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.1
json_repair==0.51.0
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
kubernetes==33.1.0
langchain-core==0.3.21
langchain-openai==0.2.10
langsmith==0.1.127
lazy-model==0.2.0
litellm==1.72.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
motor==3.4.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.108.2
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-distro==0.58b0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==24.2
pandas==2.2.2
parso==0.8.5
passlib==1.7.4
pathspec==0.12.1
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
posthog==5.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==5.9.8
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.8.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.2.2
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
redis==5.0.7
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
structlog==24.2.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.8.22
uvicorn==0.29.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
